cheque-ocr/requirements.txt
ultralytics==8.3.156       # or the version you trained with # torch==2.2.0         # CUDA image will override 2.7.1
transformers==4.52.4
pillow==9.5.0
dateparser==1.2.0            # ← correct name
fastapi==0.111.0
uvicorn[standard]==0.29.0
opencv-python-headless==4.11.0.86
